22 fix build fail (#25) (#26)
* fix location of use router on admin menu

* admin layout now server component

* change eslintrc to fix babel error
