+++ client/.eslintrc.json
@@ -1,3 +1,11 @@
 {
-  "extends": ["next/babel","next/core-web-vitals"]
+  // "extends": ["next/babel","next/core-web-vitals"]
+  "extends": [
+    "next/core-web-vitals","next"
+  ],
+  "rules": {
+    "react/no-unescaped-entities":"off",
+    "react/display-name":"off",
+    "import/no-anonymous-default-export":"off"
+  }
 }
