+++ client/package.json
@@ -19,9 +19,11 @@
     "react-dom": "^18",
     "react-firebase-hooks": "^5.1.1",
     "react-icons": "^5.1.0",
+    "typescript": "^5.4.5",
     "uuid": "^9.0.1"
   },
   "devDependencies": {
+    "@next/eslint-plugin-next": "^14.2.3",
     "@types/react": "18.3.2",
     "@types/uuid": "^9.0.8",
     "autoprefixer": "^10.0.1",
